feat: Configured Apollo Server with Helmet security middleware.
When you add `helmet` to your server, it includes a collection of 12 middleware functions. In development mode there are two of them which cause the Apollo Server GraphiQL sandbox to fail:

- crossOriginEmbedderPolicy
- contentSecurityPolicy

These rules need to be disable. Since GraphiQL is served as static, client side app which fetches GraphQL schema. This violates the default content security policy, requiring this middleware to be disabled to work.