refactor: Move logic to launch Apollo Server to Api class.
The idea is to keep our GraphQL schema definition with a version name, so in the future we can design different Schemas under different version names.

The API class encapsulates the logic to create new ApolloServices, and provides the conventions to define new GraphQL Schema versions in the project.

There is still some improvements to be made, but with this first iteration we are already in a good way to version our API specifications.