feat: Resolver for requesting Device is working.
First attempt to implement a resolver which makes use of the SDK package worked successfully.

Now we have a mutation to pair a new device to the cloud and return its token.

There are some issues with the code I need to address in upcoming PRs, but for now the logic seems to work properly.