feat: document schema, mutation and service created.
I use the `graphql-upload-ts` to enable file uploading to the server. The package adds a new Middleware for file uploading in requests, and a new `Upload` scalar.

Once this is set, we can read the stream from the `Upload` input and generate from there the `Buffer` which is then used to upload the file to reMarkable.

In the Apollo Studio, you can test this feature by adding a file with a key pointing to the corresponding input field containing the `Upload` payload, and then adding an `x-apollo-operation-name` header with the name of the operation.

In this case, if I want to upload a sample file I need to set my variables as:

```
{
  "input": {
    "name": "mydocument",
    "buffer": null
  }
}
```

And the uploaded file with the key: `input.buffer`.

There might be adjustments to be made in production, but for now this is working properly.